# resources/Wasteless_API_Spec.yml
openapi: 3.0.0
servers:
  - description: SwaggerHub API Auto Mocking
    url: >-
      https://virtserver.swaggerhub.com/matthewminish/seng302-2021-api-spec/3.0.1
info:
  description: API specification for the SENG302 2021 project team 800
  version: 6.0.0
  title: Wasteless
  contact:
    email: seng302@canterbury.ac.nz
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
security:
  - CookieAuth:
      - user
      - globalApplicationAdmin
      - defaultGlobalApplicationAdmin
paths:
...
  /users:
    post:
      tags:
        - 'U1: Registering and logging into an individual account'
        - Users
      summary: Create a new user account
      security: []
      responses:
        '201':
          $ref: '#/components/responses/AuthenticatedResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '409':
          description: Email address already in use
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        description: >-
          New user account data, id will be generated automatically by the
          server
  /users/{id}:
    get:
      tags:
        - 'U1: Registering and logging into an individual account'
        - Users
      summary: Retrieve a specific user account
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Account retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
    put:
      summary: Update a user's attributes
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: integer
      tags:
        - 'U10: Modifying Individuals'
        - Users
      responses:
        '200':
          description: User information has been updated successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          description: User not found
        '409':
          description: Email address already in use
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        description: Updated user account data
  /users/currency/{id}:
    put:
      tags:
        - 'U10: Modifying Individuals'
        - Users
      summary: Update user's currency to match with current country
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User's currency has been updated successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          description: User not found
  /user/loginCheck:
    get:
      tags:
        - 'U1: Registering and logging into an individual account'
        - Users
      summary: (Custom) checks that the user has not logged out
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      responses:
        '200':
          description: The user's session token is correct
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /users/{id}/makeadmin:
    put:
      tags:
        - 'U4: Default Global Application Admin (DGAA)'
        - Users
      summary: Promote a user account to the Global Application Admin role
      security:
        - CookieAuth:
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Action completed successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
  /users/{id}/revokeadmin:
    put:
      tags:
        - 'U4: Default Global Application Admin (DGAA)'
        - Users
      summary: Revokes the Global Application Admin role from a user account
      security:
        - CookieAuth:
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Action completed successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
        '409':
          description: >-
            Returned when the DGAA tries to revoke their own admin status, this
            is not an allowed action.
  /users/images/{imageId}/makeprimary:
    put:
      tags:
        - 'U10: Modifying Individuals'
        - Users
      summary: (Custom) set an existing image to be the new primary image for a user
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: imageId
          in: path
          description: >-
            ID of the image to set as being the primary image for the current
            user
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Primary image successfully updated
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
  /users/{id}/images:
    post:
      tags:
        - 'U10: Modifying Individuals'
        - Users
      summary: Add a new image to a user's profile
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: integer
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                filename:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: Action completed successfully
        '400':
          description: User id does not exist or there is something wrong with the file
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /users/images/{id}:
    get:
      tags:
        - 'U10: Modifying Individuals'
        - Users
        - Images
      summary: Retrieves a user image
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: ID of image being requested
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Thumbnail retrieved successfully.
        '401':
          description: Access token is missing or invalid
        '406':
          description: >-
            The requested route does exist (so not a 404) but some part of the
            request is not acceptable, for example trying to access a resource
            by an ID that does not exist.
  /users/images/{id}/thumbnail:
    get:
      tags:
        - 'U10: Modifying Individuals'
        - Users
        - Images
      summary: Retrieves a thumbnail image for a user
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: ID of thumbnail image being requested
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Thumbnail retrieved successfully.
        '401':
          description: Access token is missing or invalid
        '406':
          description: >-
            The requested route does exist (so not a 404) but some part of the
            request is not acceptable, for example trying to access a resource
            by an ID that does not exist.
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  responses:
    BadRequestError:
      description: >-
        There was some error with the data supplied by the user, appropriate
        error message(s) should be shown to user
    UnauthorizedError:
      description: Access token is missing or invalid
    ForbiddenError:
      description: The user does not have permission to perform the requested action
    NotAcceptableError:
      description: >-
        The requested route does exist (so not a 404) but some part of the
        request is not acceptable, for example trying to access a resource by an
        ID that does not exist.
    AuthenticatedResponse:
      description: >
        Response returned to client when they have performed an action to gain
        authentication (registering or logging in). This response includes a
        session token that the client can use in future API requests to
        authenticate itself. This session token is set as a cookie with name
        'JSESSIONID', and will need to be included in subsequent requests to the
        server.
      content:
        application/json:
          schema:
            type: object
            properties:
              userId:
                type: integer
                description: The ID of the user that has just been authenticated
                example: 100
      headers:
        Set-Cookie:
          schema:
            type: string
            example: JSESSIONID=abcde12345; Path=/; HttpOnly
  schemas:
...
    Business:
      type: object
      description: Business account entity
      required:
        - name
        - address
        - businessType
      properties:
        id:
          type: integer
          example: 100
          readOnly: true
          description: Automatically generated and assigned by the server
        administrators:
          type: array
          items:
            $ref: '#/components/schemas/User'
          readOnly: true
        primaryAdministratorId:
          type: integer
          example: 20
          description: >-
            The ID of the 'primary' administrator of the business. This is
            automatically given to the creator of the business.
        name:
          type: string
          example: Lumbridge General Store
        description:
          type: string
          example: A one-stop shop for all your adventuring needs
        address:
          $ref: '#/components/schemas/Address'
        businessType:
          type: string
          enum:
            - Accommodation and Food Services
            - Retail Trade
            - Charitable organisation
            - Non-profit organisation
        created:
          type: string
          format: date-time
          example: '2020-07-14T14:52:00Z'
          readOnly: true
          description: Date and time when the account was created
    Address:
      type: object
      description: Entity representing some physical address
      required:
        - country
      properties:
        streetNumber:
          type: string
          example: 3/24
          writeOnly: true
        streetName:
          type: string
          example: Ilam Road
          writeOnly: true
        suburb:
          type: string
          example: Upper Riccarton
          writeOnly: true
        city:
          type: string
          example: Christchurch
        region:
          type: string
          example: Canterbury
        country:
          type: string
          example: New Zealand
        postcode:
          type: string
          example: '90210'
          writeOnly: true
    User:
      type: object
      description: Basic user entity
      required:
        - firstName
        - lastName
        - email
        - dateOfBirth
        - homeAddress
        - password
      properties:
        id:
          type: integer
          example: 100
          readOnly: true
          description: Automatically generated and assigned by the server
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Smith
        middleName:
          type: string
          example: Hector
        nickname:
          type: string
          example: Jonny
        bio:
          type: string
          example: Likes long walks on the beach
        email:
          type: string
          format: email
          example: johnsmith99@gmail.com
          writeOnly: true
        primaryImageId:
          type: integer
          example: 1
          description: >-
            The ID of the 'primary' image chosen by the user. By default it is
            the first image the use uploads.
        countryForCurrency:
          type: string
          description: >-
            The country in which the currency is used, if not initialized then
            default to New Zealand
          example: New Zealand
        dateOfBirth:
          type: string
          format: date
          example: '1999-04-27'
          writeOnly: true
        phoneNumber:
          type: string
          example: +64 3 555 0129
          writeOnly: true
        homeAddress:
          $ref: '#/components/schemas/Address'
        password:
          type: string
          example: 1337-H%nt3r2
          format: password
          writeOnly: true
          description: Should not be stored in plain text (i.e hash it)
        created:
          type: string
          format: date-time
          example: '2020-07-14T14:32:00Z'
          readOnly: true
          description: Date and time when the account was created
        role:
          type: string
          readOnly: true
          description: >-
            (Implemented in U4): This property should only be shown to Global
            Application Admins
          enum:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
        businessesAdministered:
          type: array
          readOnly: true
          description: >-
            (Implemented in U6): A list of businesses for which the user is an
            administrator
          items:
            $ref: '#/components/schemas/Business'
...